refactor: move notification related api to its own file
